p1: drivers: power: force charging status if usb cable is connected
although the usb cable will be recognized during charging process
but it won't get power supply enought from computer if the tab screen
is on for at least 6 minutes.

the main reason of this patch is the system notification
for charging process from usb cable type.

the improper TA cable will not get charging status as
it is for default reason.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: I1ab0850cbfdfcdf81546e857528c385794fa8203